Ensured no tokens in A are the same as in B in letter for letter #618
Made sure that A and B have no overlap in tokens for letter for letter in baconian. Also made a method for buildset so you can get the non-bitmap form of the built set if necessary, while keeping original functionality by making it a seperate method (could maybe have overloaded it instead of new method? maybe consider doing so in the future)