Update cipheraristocratencoder.ts
To address #658. I kept running into an uncaught runtime error when attempting to test - something about the suggested scores with the Aristo encoder is doing that - so I'd say let's push it out and we can tweak it if we find they are too low.  The goal would be to have these in the low-mid 400s, I think, and they are showing up in the mid 600s.